+++ src/fits_read/error.rs
@@ -2,9 +2,9 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-Errors associated with reading in fits files.
- */
+//! Errors associated with reading in fits files.
+
+use std::path::PathBuf;
 
 use thiserror::Error;
 
@@ -14,7 +14,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\nCouldn't open {fits_filename}: {fits_error}")]
     Open {
         fits_error: fitsio::errors::Error,
-        fits_filename: String,
+        fits_filename: PathBuf,
         source_file: &'static str,
         source_line: u32,
     },
@@ -23,7 +23,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: Couldn't find key {key}")]
     MissingKey {
         key: String,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -32,7 +32,7 @@ pub enum FitsError {
     /// Error describing a HDU that couldn't be used as an image (e.g. `HduInfo::ImageInfo`).
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: Tried to use as an image, but not an image")]
     NotImage {
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -42,7 +42,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: Couldn't read a long string from {key}")]
     LongString {
         key: String,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -52,7 +52,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\n{fits_filename} HDU {hdu_num}: {fits_error}")]
     Fitsio {
         fits_error: fitsio::errors::Error,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,
@@ -62,7 +62,7 @@ pub enum FitsError {
     #[error("{source_file}:{source_line}\nCouldn't parse {key} in {fits_filename} HDU {hdu_num}")]
     Parse {
         key: String,
-        fits_filename: String,
+        fits_filename: PathBuf,
         hdu_num: usize,
         source_file: &'static str,
         source_line: u32,