+++ src/metafits_context/mod.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-The main interface to MWA data.
- */
+//! The main interface to MWA data.
+
 use std::fmt;
 use std::path::Path;
 
@@ -27,7 +26,7 @@ mod test;
 /// Enum for all of the known variants of file format based on Correlator version
 ///
 #[repr(C)]
-#[derive(Debug, PartialEq, Clone, Copy)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 pub enum MWAVersion {
     /// MWA correlator (v1.0), having data files without any batch numbers.
     CorrOldLegacy = 1,
@@ -106,7 +105,7 @@ impl fmt::Display for VisPol {
 }
 
 #[repr(C)]
-#[derive(Debug, PartialEq, Clone, Copy, FromPrimitive)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, FromPrimitive)]
 pub enum GeometricDelaysApplied {
     No = 0,
     Zenith = 1,
@@ -156,7 +155,7 @@ impl std::str::FromStr for GeometricDelaysApplied {
 }
 
 #[repr(C)]
-#[derive(Debug, PartialEq, Clone, Copy, FromPrimitive)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, FromPrimitive)]
 pub enum CableDelaysApplied {
     NoCableDelaysApplied = 0,
     CableAndRecClock = 1,
@@ -206,7 +205,7 @@ impl std::str::FromStr for CableDelaysApplied {
 }
 
 #[repr(C)]
-#[derive(Debug, PartialEq, Clone, Copy)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
 #[allow(non_camel_case_types, clippy::upper_case_acronyms)]
 pub enum MWAMode {
     No_Capture = 0,