+++ src/metafits_context/test.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-Unit tests for metafits context
-*/
+//! Unit tests for metafits context
+
 use std::str::FromStr;
 
 #[cfg(test)]
@@ -16,7 +15,7 @@ fn test_metafits_context_new_invalid() {
     let metafits_filename = "invalid.metafits";
 
     // No gpubox files provided
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::CorrMWAXv2));
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::CorrMWAXv2));
 
     assert!(context.is_err());
 }
@@ -30,7 +29,7 @@ fn test_metafits_context_new_vcs_legacy_valid() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSLegacyRecombined))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSLegacyRecombined))
         .expect("Failed to create MetafitsContext");
 
     // rf_inputs:                [Tile104Y, ..., Tile055X],
@@ -63,7 +62,7 @@ fn test_metafits_context_new_corrlegacy_valid() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::CorrLegacy))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::CorrLegacy))
         .expect("Failed to create MetafitsContext");
 
     // Test the properties of the context object match what we expect
@@ -268,7 +267,7 @@ fn test_metafits_context_new_corrmwaxv2_valid() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::CorrMWAXv2))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::CorrMWAXv2))
         .expect("Failed to create MetafitsContext");
 
     // Test the properties of the context object match what we expect
@@ -292,7 +291,7 @@ fn test_metafits_context_new_vcsmwax2_valid() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSMWAXv2))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSMWAXv2))
         .expect("Failed to create MetafitsContext");
 
     // Test the properties of the context object match what we expect
@@ -326,7 +325,7 @@ fn test_populate_expected_timesteps() {
 
     for mwa_version in mwa_versions {
         // Open a context and load in a test metafits
-        let result = MetafitsContext::new_internal(&metafits_filename);
+        let result = MetafitsContext::new_internal(metafits_filename);
 
         assert!(result.is_ok());
 
@@ -367,7 +366,7 @@ fn test_populate_expected_coarse_channels_legacy() {
 
     for mwa_version in mwa_versions {
         // Open a context and load in a test metafits
-        let result = MetafitsContext::new_internal(&metafits_filename);
+        let result = MetafitsContext::new_internal(metafits_filename);
 
         assert!(result.is_ok());
 
@@ -410,7 +409,7 @@ fn test_populate_expected_coarse_channels_corr_mwaxv2() {
 
     for mwa_version in mwa_versions {
         // Open a context and load in a test metafits
-        let result = MetafitsContext::new_internal(&metafits_filename);
+        let result = MetafitsContext::new_internal(metafits_filename);
 
         assert!(result.is_ok());
 
@@ -450,7 +449,7 @@ fn test_metafits_context_new_guess_version() {
     let metafits_filename = "test_files/1101503312_1_timestep/1101503312.metafits";
 
     // Open a context and load in a test metafits
-    let result = MetafitsContext::new(&metafits_filename, None);
+    let result = MetafitsContext::new(metafits_filename, None);
     assert!(result.is_ok());
 
     let context = result.unwrap();
@@ -463,7 +462,7 @@ fn test_generate_expected_volt_filename_legacy_vcs() {
     let metafits_filename = "test_files/1101503312_1_timestep/1101503312.metafits";
 
     // Open a context and load in a test metafits
-    let result = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSLegacyRecombined));
+    let result = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSLegacyRecombined));
     assert!(result.is_ok());
 
     let context = result.unwrap();
@@ -479,7 +478,7 @@ fn test_generate_expected_volt_filename_mwax_vcs() {
     let metafits_filename = "test_files/1101503312_1_timestep/1101503312.metafits";
 
     // Open a context and load in a test metafits
-    let result = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSMWAXv2));
+    let result = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSMWAXv2));
     assert!(result.is_ok());
 
     let context = result.unwrap();
@@ -495,7 +494,7 @@ fn test_generate_expected_volt_filename_invalid_timestep() {
     let metafits_filename = "test_files/1101503312_1_timestep/1101503312.metafits";
 
     // Open a context and load in a test metafits
-    let result = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSLegacyRecombined));
+    let result = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSLegacyRecombined));
     assert!(result.is_ok());
 
     let context = result.unwrap();
@@ -509,7 +508,7 @@ fn test_generate_expected_volt_filename_invalid_coarse_chan() {
     let metafits_filename = "test_files/1101503312_1_timestep/1101503312.metafits";
 
     // Open a context and load in a test metafits
-    let result = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSLegacyRecombined));
+    let result = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSLegacyRecombined));
     assert!(result.is_ok());
 
     let context = result.unwrap();