+++ src/misc/mod.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-General helper/utility methods
-*/
+//! General helper/utility methods
+
 use crate::antenna;
 use std::{mem, slice};