+++ src/voltage_files/test.rs
@@ -2,9 +2,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-/*!
-Unit tests for voltage file metadata
-*/
+//! Unit tests for voltage file metadata
+
 #[cfg(test)]
 use super::*;
 use std::fs::File;
@@ -553,7 +552,7 @@ fn test_examine_voltage_files_valid() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSMWAXv2))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSMWAXv2))
         .expect("Failed to create MetafitsContext");
 
     // Create a temp dir for the temp files
@@ -593,7 +592,7 @@ fn test_examine_voltage_files_error_mismatched_sizes() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSMWAXv2))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSMWAXv2))
         .expect("Failed to create MetafitsContext");
 
     // Create a temp dir for the temp files
@@ -641,7 +640,7 @@ fn test_examine_voltage_files_error_gpstime_gaps() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSMWAXv2))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSMWAXv2))
         .expect("Failed to create MetafitsContext");
 
     // Create a temp dir for the temp files
@@ -686,7 +685,7 @@ fn test_examine_voltage_files_error_file_not_found() {
     // Read the observation using mwalib
     //
     // Open a context and load in a test metafits
-    let context = MetafitsContext::new(&metafits_filename, Some(MWAVersion::VCSMWAXv2))
+    let context = MetafitsContext::new(metafits_filename, Some(MWAVersion::VCSMWAXv2))
         .expect("Failed to create MetafitsContext");
 
     // Populate vector of filenames